添加最短路径实现
添加下列最短路径实现
广度搜索
Floyd
dijstra

+++ 961数据结构/Graph/GraphEdge.cs
@@ -8,7 +8,7 @@ namespace _961数据结构.Graph
     {
         public GraphNode StartNode { get; set; }
         public GraphNode EndNode { get; set; }
-        public int Weights { get; set; }
+        public double Weights { get; set; }
 
         public int CompareTo(object obj)
         {
